Management of zoom slider in navigation widget (top left of GUI)

+++ trunk/workflow/inc/emamainwindow.h
@@ -222,7 +222,7 @@ private slots:
 
 	void slot_mainGroupBox_resizeEvent(QResizeEvent *);
 	//
-	void on_globalNavImageWidget_signalZoomOn(int, int, int);
+	void on_globalNavImageWidget_signalZoomOn(int, int, float);
 
 	//void on_mainDisplayWidget_signalGreyscaleToggled(bool);
 	void slot_mainDisplayWidget_signalZoomRect(QRect cropRect);
